VOL-1604 Implementing storage of MIB entities in etcd.
Updated tests to work and pulled a test update from https://gerrit.opencord.org/#/c/12815/. The functional changes were merged some time ago, but the test was not.
Updated requirements to match those in voltha-openonu-adapter. This fixed an issue where the test would fail on importing the etcd database class
Change-Id: I13787dbf740020eb2d7f2c6498f9eafe1b461cba