VOL-2171: MDS values calculated incorrectly
G.984 and G.988 both specify the MDS increment
must not include 0.  Fix that.

Also when reconciliation is complete use the successes
value for setting the new mds on the onu.  This value
is more reliable as the old mds increment was based on
the previous sync.

Also this forces the mds set to not be based on a query thats
subject to the event bus on_set_response queue, that may still be
catching up recording the updates to the database.  Setting the MDS from
within the on_set_response using what was actually sent allows all other
events to be processed in order and ensures that whats set on the onu matches
what is set in openomci database.  Before the mds was queried and set too
soon and updates came in later and the values did not match.

Also provide a way to query vendor_id

Change-Id: I0883b25ae59ef5eef409cd76f7c0e1012e8388fe